backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
openai
mcp
httpx
google-generativeai
email-validator
